requirements.txt
kafka-python
pandas
scikit-learn
openai
streamlit
streamlit-autorefresh
